[UPDATE]: Renamed library feature to learning and organized file structure
